requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements/requirements.in
#
configparser==7.2.0
    # via -r requirements/requirements.in
kafka-python==2.3.0
    # via -r requirements/requirements.in
numpy==2.2.6
    # via pandas
pandas==2.3.3
    # via -r requirements/requirements.in
python-dateutil==2.9.0.post0
    # via pandas
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via -r requirements/requirements.in
six==1.17.0
    # via python-dateutil
tzdata==2025.3
    # via pandas
